update export script, embed rstIncludeToMdx functions
